Main Menu, Controls Menu added and working
Main Menu, Controls Menu added and working.
Exit game from main menu and pause menu.

NOTE: Must "launch" game in order to use esc for pause menu.

NOTE: In order to use main menu from play in editor, you must open the Main Menu Map.
Main menu will not open automatically yet. this needs to be changed in the settings. I thought it was fine to leave it for now as we work on the game. To avoid annoyance. This will need to be changed before packaging.