MATs Added
Lava material added, heat distortion added. Light adhered to cargo so that when it spawns, the shadow is more distinct. Light disappears after cargo falls.

With the magma material added, it was hard to see where the cargo was falling. Needed to strengthen the shadow.

Projectile spawns in middle of scene after 2 seconds. This will go away tomorrow. I left it this way to help me remember where I left off.